make corrections for correct egg output
egg output was wrong, not multiplied by time step. To correct this, we edited the egg output function, egg uptake function, which was multiplied by time step but shouldn't be. also cercariae death was edited to make it consistent across time steps and prevalence calculation was also edited to reflect the fact that eggs are now multiplied by the time step and hence the sampled eggs on a given day now have to be divided by time step